Trials/react/image-generator/text-to-image-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pillow
huggingface_hub
